requirements.txt
google-genai>=0.7.0